feat: add createpost and createcomment
